requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.5.0.5
Django==3.0
importlib-metadata==1.6.0
kombu==4.6.8
pytz==2019.3
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
celery==4.4.2
Pillow==7.1.1
boto3==1.12.40
ImageHash==4.0